feat: 판별 결과를 첨부·메일 상태에 반영 (#15)
* feat: 판별 결과를 첨부·메일 상태에 반영

분석 파이프라인은 들어와 있었지만 양쪽 끝이 끊겨 있었다. enqueue 를
부르는 운영 코드가 없어 큐에 아무것도 들어가지 않았고, DecisionResponse
를 받는 리스너가 없어 판정이 DB 에 닿지 않았다. 첨부는 전부 PENDING 에
머물렀다.

AttachmentScanBridge 가 두 방향을 모두 맡는다. 첨부 도메인에서 upstage
를 호출하는 곳은 여기 하나뿐이고 타입 변환도 여기서 한다.

- 등록 커밋 이후에 검사를 요청한다. 트랜잭션 안에서 보내면 워커가 아직
  커밋되지 않은 행을 찾다가 결과를 버린다
- 큐 키로 메일 id 가 아니라 첨부 id 를 쓴다. MailStore 가 mailId 로
  덮어써서, 같은 메일에 첨부를 연달아 붙이면 먼저 온 요청이 사라진다.
  결과는 storageKey 로 돌아오므로 반영에는 영향이 없다
- 해시로 이전 판정을 물려받았으면 다시 분석하지 않는다
- upstage 의 ALLOW/REVIEW 를 첨부 도메인의 ALLOWED/BLOCKED 로 옮긴다

발송 판단도 판정을 본다. 차단 판정 첨부가 있으면 사내 발송이라도 보류로
가고, 아직 검사 중인 첨부가 있으면 409 로 막는다. 검사가 끝나지 않은
파일이 그대로 나가던 구멍이었다.

PR #10 이 남긴 EmailService 의 미사용 import 두 개도 정리했다.

* fix: 테스트에서 ScanStatus 이름 충돌 해소

* fix: 분석이 실패해도 판정을 발행한다

MailProcessor 에 예외 처리가 없어서, Upstage 호출이 실패하면
publishDecision 까지 못 가고 그대로 빠져나갔다. 그러면 첨부가 PENDING 에
영구히 남아 화면은 "검사 중"을 무한히 돌고 발송은 계속 409 로 막힌다.
사용자가 그 메일을 영영 보낼 수 없게 된다.

첨부 단위와 그 바깥(정책 조회 등) 양쪽에서 예외를 잡고, 검사하지 못한
첨부는 REVIEW 로 넘긴다. 통과가 아니라 보류다 — 못 본 파일을 그냥
내보내는 것이 이 서비스에서 제일 나쁜 실패다.

첨부 하나가 터져도 같은 메일의 나머지는 정상 판정을 받는다.

* refactor: upstage 식별자를 Long 으로 통일하고 큐 키를 요청 단위로 분리

* fix: 테스트 컴파일 오류 (Long 리터럴, Mail 생성자 가시성)

+++ src/main/java/com/shhdoc/upstage/Gateway.java
@@ -40,7 +40,7 @@ public interface Gateway {
      * @param companyId 조회 대상 기업 식별자
      * @return 미완료 메일들의 처리 상태 목록. 미완료 건이 없으면 빈 리스트를 반환합니다.
      */
-    List<MailStatusResponse> getStatus(Integer companyId);
+    List<MailStatusResponse> getStatus(Long companyId);
 
     /**
      * ShhDoc이 처리를 완료한 메일의 판단 결과(첨부파일별 ALLOW/REVIEW)를
