+++ src/test/java/com/shhdoc/upstage/GatewayEndToEndTest.java
@@ -54,28 +54,28 @@ class GatewayEndToEndTest {
     private ApplicationEvents events;
 
     private record DocumentCase(
-            Integer mailId, String fileName, String storageKey, String recipientAddress, ScanStatus expectedStatus) {
+            Long mailId, String fileName, String storageKey, String recipientAddress, ScanStatus expectedStatus) {
     }
 
     private static final List<DocumentCase> CASES = List.of(
             // 급여명세서, 외부발송: mock 정책상 payslip+외부(미분류) 룰 → REVIEW
-            new DocumentCase(1001, "sample-payslip.pdf", "storage-key-payslip-ext", EXTERNAL_RECIPIENT, ScanStatus.REVIEW),
+            new DocumentCase(1001L, "sample-payslip.pdf", "storage-key-payslip-ext", EXTERNAL_RECIPIENT, ScanStatus.REVIEW),
             // 급여명세서, 내부발송(같은 도메인): payslip+internal 룰 → ALLOW
-            new DocumentCase(1005, "sample-payslip.pdf", "storage-key-payslip-int", INTERNAL_RECIPIENT, ScanStatus.ALLOW),
+            new DocumentCase(1005L, "sample-payslip.pdf", "storage-key-payslip-int", INTERNAL_RECIPIENT, ScanStatus.ALLOW),
             // 계약서, 외부발송: payslip 전용 룰 안 걸림 → 와일드카드 폴백 → ALLOW
-            new DocumentCase(1002, "sample-contract.pdf", "storage-key-contract", EXTERNAL_RECIPIENT, ScanStatus.ALLOW),
+            new DocumentCase(1002L, "sample-contract.pdf", "storage-key-contract", EXTERNAL_RECIPIENT, ScanStatus.ALLOW),
             // 공지문, 외부발송: 민감정보 없음 → ALLOW (false positive 안 나는지 검증)
-            new DocumentCase(1003, "sample-notice.pdf", "storage-key-notice", EXTERNAL_RECIPIENT, ScanStatus.ALLOW),
+            new DocumentCase(1003L, "sample-notice.pdf", "storage-key-notice", EXTERNAL_RECIPIENT, ScanStatus.ALLOW),
             // 사업계획서, 외부발송: 개인정보 없이 재무+대외비만 → ALLOW (payslip과 다른 신호조합)
-            new DocumentCase(1004, "sample-business-plan.pdf", "storage-key-business-plan", EXTERNAL_RECIPIENT, ScanStatus.ALLOW)
+            new DocumentCase(1004L, "sample-business-plan.pdf", "storage-key-business-plan", EXTERNAL_RECIPIENT, ScanStatus.ALLOW)
     );
 
     @Test
     void enqueue부터_publishDecision까지_5건을_동시에_처리한다() {
         for (DocumentCase testCase : CASES) {
-            Attachment attachment = new Attachment(testCase.fileName(), 1, testCase.storageKey(), "fake-hash");
+            Attachment attachment = new Attachment(testCase.fileName(), 1L, testCase.storageKey(), "fake-hash");
             MailRequest request = new MailRequest(
-                    testCase.mailId(), 1, SENDER, 1, "제목", "본문",
+                    testCase.mailId(), 1L, SENDER, 1L, "제목", "본문",
                     List.of(new Recipient(testCase.recipientAddress())),
                     List.of(attachment)
             );
